29-JUN-2023

+++ devcamper-api/controllers/courses.js
@@ -45,6 +45,7 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 
 exports.addCourse = asyncHandler(async (req, res, next) => {
   req.body.bootcamp = req.params.bootcampId;
+  req.body.user = req.user.id;
 
   const bootcamp = await Bootcamp.findById(req.params.bootcampId);
 
@@ -54,6 +55,17 @@ exports.addCourse = asyncHandler(async (req, res, next) => {
       404
     );
   }
+
+  // Make sure user is bootcamp owner
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to add a course to bootcamp ${bootcamp._id}`,
+        401
+      )
+    );
+  }
+
   const course = await Course.create(req.body);
   res.status(201).json({ success: true, data: course });
 });
@@ -71,6 +83,17 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
       404
     );
   }
+
+  // Make sure user is course owner
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to update course to bootcamp ${course._id}`,
+        401
+      )
+    );
+  }
+
   course = await Course.findById(req.params.id, req.body, {
     new: true,
     runValidators: true,
@@ -91,6 +114,17 @@ exports.deleteCourse = asyncHandler(async (req, res, next) => {
       404
     );
   }
+
+  // Make sure user is course owner
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to delete course to bootcamp ${course._id}`,
+        401
+      )
+    );
+  }
+
   await course.deleteOne();
   res.status(200).json({ success: true, data: course });
 });
